requirements.txt
# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.20.0

# 计算机视觉
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# 机器学习与数据处理
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# 聚类算法
hdbscan>=0.8.29
umap-learn>=0.5.3

# 向量数据库
faiss-cpu>=1.7.4
pymilvus>=2.3.0

# 多模态模型
clip-by-openai>=1.0
open-clip-torch>=2.20.0

# Web框架
fastapi>=0.100.0
uvicorn>=0.23.0
flask>=2.3.0

# 前端相关
streamlit>=1.25.0

# 数据处理
pydantic>=2.0.0
python-multipart>=0.0.6

# 图像处理
imageio>=2.31.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 工具库
tqdm>=4.65.0
requests>=2.31.0
python-dotenv>=1.0.0

# 性能优化
onnx>=1.14.0
onnxruntime>=1.15.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0

# 代码质量
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0 